create folders structure to hold model output.

The folder structure is dependent on the number and type of scenarios selected. Within each scenario are subfolder to store model output dat and pin files and/or plots. Additional folders are also created ofr summary statistics and plots.

create_folder_setup(
  rootFolder,
  hcrTypes = c("Fixed", "Ramp"),
  hcrLevels = c("Complex", "HTSpecies", "LTSpecies"),
  exploitationRates,
  create = T
)

Arguments

rootFolder
Character string: name of folder to hold all output from a single set of parameters

hcrTypes
Character vector: Names for the harvest control rule types (Default = c("Fixed, "Ramp"))

hcrLevels
Character vector: Names for the levels that the harvest control rules are applied to. (Default = c("Complex","HTSpecies","LTSpecies"))

exploitationRates
Numeric vector: Values for the maximum exploitation rate in any scenario.

create
Boolean. To create folders on machine (TRUE), to retrun structure only (F)

Value

A Character matrix (3 columns):

folderStructure

Column 1 = hcrType, Column 2 = hcrLevels, column 3 = exploitationRate