Process all runs for a given scenario

A scenario is defined a a harvest control rule with a maximum exploitation rate. Each scenario is run a number of times (nSims). All of the runs for a scenario are processed. Specifics of processing are expalined in section below

process_single_scenario(
  modelOutput,
  indices,
  revenueData,
  convertBiomassTolbs = 2204.62,
  nLastYrs = 20
)

Arguments

modelOutput
List (length = nSims) content of files created by running Hydra

indices
Character vector. Names of the indices to process

revenueData
Data frame. External revenue data (names = character vector of species names, ppp = price ($) per pound)

convertBiomassTolbs
Numeric scalar. 1 metric tonne = 2204.62 lbs (default)

nLastYrs
Numeric scalar. Number of trailing years in each run to take mean over. Default = 20 (Last 20 yrs of sim)

Value

A list containing the following items:

species_bio

matrix (nYears x nSpecies) of median biomass (over nSims)

species_catch

matrix (nYears x nSpecies) of median catch (over nsims)

guild_bio

matrix (nYears n nGuilds) of median biomass by guild (over nsims)

guild_catch

matrix (nYears x nGuilds) of median catch by guild (over nsims)

dfI

data frame. Each column represents an index. Each row represent a year. see section below

catch_runs

matrix (nGuilds x nruns). Mean of the last x years of each run

Processing specifics

species_bio, species_catch, guild_bio, guild_bio are calculated from ("avByr","est_catch_biomass","est_survey_guild_biomass","est_catch_guild_biomass") by taking the median value over the nSims model runs.

Each index in indices is processed and returned in dfI

For indices that are not species specific (Diversity indices, Large fish indices) the median value over all nSims model runs are used at each time point

For indices that are species dependent (avByr, est_catch_biomass). The median is calculated (over all nSims) for each species at each time point. The mean of the species medians are then used to represent the returned index value

For indices that are species specific that involve status (species or guild status) the index is calculated as the total number of times a species (or guild) falls below the minimum threshold specified in the model

Revenue is calculated from "est_catch_biomass". Biomass is converted to lbs. The median weight for each species in each year is multiplied by the price per pound (currently 2012 price)